+++ src/data/general_store.cpp
@@ -3,21 +3,21 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 #include "general_store.h"
 
- volatile double vbat_global = 0;
- volatile double teng_global = 0;
+ volatile float vbat_global = 0;
+ volatile float teng_global = 0;
 
- void vbatGlobalRead(double& out) {
+ void vbatGlobalRead(float& out) {
      out = vbat_global;
  }
 
- void vbatGlobalWrite(const double& in) {
+ void vbatGlobalWrite(const float& in) {
      vbat_global = in;
  }
 
- void tengGlobalRead(double& out) {
+ void tengGlobalRead(float& out) {
      out = teng_global;
  }
 
- void tengGlobalWrite(const double& in) {
+ void tengGlobalWrite(const float& in) {
      teng_global = in;
  }
